requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
torch>=2.0.0
torchvision>=0.15.0
